+++ src/components/pages/articles/ArticleWritingPage/index.tsx
@@ -1,11 +1,12 @@
+/** @jsxImportSource @emotion/react */
+
 import { useParams } from 'react-router-dom';
-import styled from 'styled-components';
 import qs from 'qs';
 
-import { Container } from '@root/components/atoms/global';
 import { isEmpty } from '@root/common/utils';
-import WritingContent from './WritingContent';
-import Board from '../common/Board';
+import WritingContent from './components/WritingContent';
+import Board from '@components/pages/articles/common/Board';
+import { css } from '@emotion/react';
 
 const ArticleWritingPage = () => {
   const { id } = useParams();
@@ -15,55 +16,29 @@ const ArticleWritingPage = () => {
 
   return (
     <Board categoryId={categoryIdNumber}>
-      <WritingBlock>
-        <div className="block writing_block">
+      <main css={mainWritingBlock}>
+        <div css={innerWritingBlock}>
           <WritingContent categoryId={categoryIdNumber} articleId={articleId} />
         </div>
-      </WritingBlock>
+      </main>
     </Board>
   );
 };
 
-const WritingBlock = styled(Container)`
+const mainWritingBlock = css`
   display: flex;
   flex-direction: row;
+
   margin-top: 1.5rem;
+  margin: auto;
 
-  & > .block {
-    margin: 0 0.8rem;
-    width: 100%;
-  }
-  .category_block {
-    width: 12rem;
-    min-width: 12rem;
-  }
-  .writing_block {
-    max-width: calc(100% - 15.2rem - 9rem - 1.6rem);
-    & > div {
-      margin-bottom: 1.5rem;
-    }
-  }
+  width: 100%;
+  max-width: 1100px;
+`;
 
-  @media screen and (max-width: 1020px) {
-    .writing_block {
-      max-width: calc(100% - 15.2rem);
-      & > div {
-        margin-bottom: 1.5rem;
-      }
-    }
-  }
-
-  ${props => props.theme.mobileSize} {
-    margin-top: 0;
-    .category_block {
-      display: none;
-    }
-    .writing_block {
-      width: 100%;
-      max-width: 100%;
-      margin: 0;
-    }
-  }
+const innerWritingBlock = css`
+  margin: 0 0.8rem;
+  width: 100%;
 `;
 
 export default ArticleWritingPage;
